Updated the ERD. Move the files to a subfolder. Connect the main file with the database using API
