+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "smartseat",
+            "cwd": "frontend\\smartseat",
+            "request": "launch",
+            "type": "dart",
+            "program": "lib/screens/main.dart",
+        },
+        {
+            "name": "smartseat (profile mode)",
+            "cwd": "frontend\\smartseat",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "profile"
+        },
+        {
+            "name": "smartseat (release mode)",
+            "cwd": "frontend\\smartseat",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "release"
+        }
+    ]
+}
